create useSelectedDate hooks
